backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation & settings
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1

# HTTP client (Spotify API 호출)
httpx==0.27.0

# Rate limiting
slowapi==0.1.9
